+++ Autopilot/AttitudeManager/InputCapture.h
@@ -35,6 +35,6 @@ void initIC(unsigned char initIC);
  * that is disconnected will have a value of 0.
  * @return Array containing all the channel values
  */
-unsigned int* getICValues(unsigned int sys_time);
+unsigned int* getICValues(unsigned long int sys_time);
 
 #endif
